clean a bit the logging
