+++ app/database/db.go
@@ -2,12 +2,13 @@ package database
 
 import (
 	"fmt"
+	"log"
+	"os"
 	"reflect"
 	"strings"
 
 	"github.com/jinzhu/gorm"
 
-	"github.com/France-ioi/AlgoreaBackend/app/logging"
 	"github.com/France-ioi/AlgoreaBackend/app/types"
 )
 
@@ -30,12 +31,24 @@ func Open(dsnConfig string) (*DB, error) {
 	dbConn, err = gorm.Open(driverName, dsnConfig)
 	dbConn.LogMode(true)
 
-	// setup logging
-	if logging.Logger != nil {
-		dbConn.SetLogger(logging.Logger.WithField("module", "database"))
-	}
-
 	return newDB(dbConn), err
+}
+
+// DBLogger is the logger type for the DB logs
+type DBLogger interface {
+	Print(v ...interface{})
+}
+
+// SetLogger sets the logger to use for db and the verbosity level.
+// Pass `nil` as logger to disable logging (only errors to stdout)
+func (conn *DB) SetLogger(logger DBLogger) {
+	if logger == nil {
+		conn.db.LogMode(false)
+		conn.db.SetLogger(gorm.Logger{LogWriter: log.New(os.Stdout, "\r\n", 0)})
+	} else {
+		conn.db.LogMode(true)
+		conn.db.SetLogger(logger)
+	}
 }
 
 func (conn *DB) inTransaction(txFunc func(*DB) error) (err error) {
