--- app/logging/logger.go
@@ -1,157 +1,0 @@
-package logging
-
-// MIT License
-
-// Copyright (c) 2017 Dikton Haxhijaj
-
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
-import (
-	"fmt"
-	"log"
-	"net/http"
-	"time"
-
-	"github.com/go-chi/chi/middleware"
-	"github.com/sirupsen/logrus"
-
-	"github.com/France-ioi/AlgoreaBackend/app/config"
-)
-
-var (
-	// Logger is a configured logrus.Logger.
-	Logger *logrus.Logger
-)
-
-// StructuredLogger is a structured logrus Logger.
-type StructuredLogger struct {
-	Logger *logrus.Logger
-}
-
-// New creates and configures a new logrus Logger.
-func New(conf config.Logging) *logrus.Logger {
-	var err error
-
-	Logger = logrus.New()
-	if conf.TextLogging {
-		Logger.Formatter = &logrus.TextFormatter{
-			DisableTimestamp: true,
-		}
-	} else {
-		Logger.Formatter = &logrus.JSONFormatter{}
-	}
-
-	level := conf.LogLevel
-	if level == "" {
-		level = "error"
-	}
-	var l logrus.Level
-	l, err = logrus.ParseLevel(level)
-	if err != nil {
-		log.Fatal(err)
-	}
-	Logger.Level = l
-	return Logger
-}
-
-// NewStructuredLogger implements a custom structured logrus Logger.
-func NewStructuredLogger(logger *logrus.Logger) func(next http.Handler) http.Handler {
-	return middleware.RequestLogger(&StructuredLogger{logger})
-}
-
-// NewLogEntry sets default request log fields.
-func (l *StructuredLogger) NewLogEntry(r *http.Request) middleware.LogEntry {
-	entry := &StructuredLoggerEntry{Logger: logrus.NewEntry(l.Logger)}
-	logFields := logrus.Fields{}
-
-	logFields["ts"] = time.Now().UTC().Format(time.RFC1123)
-
-	if reqID := middleware.GetReqID(r.Context()); reqID != "" {
-		logFields["req_id"] = reqID
-	}
-
-	scheme := "http"
-	if r.TLS != nil {
-		scheme = "https"
-	}
-
-	logFields["http_scheme"] = scheme
-	logFields["http_proto"] = r.Proto
-	logFields["http_method"] = r.Method
-
-	logFields["remote_addr"] = r.RemoteAddr
-	logFields["user_agent"] = r.UserAgent()
-
-	logFields["uri"] = fmt.Sprintf("%s://%s%s", scheme, r.Host, r.RequestURI)
-
-	entry.Logger = entry.Logger.WithFields(logFields)
-
-	entry.Logger.Infoln("request started")
-
-	return entry
-}
-
-// StructuredLoggerEntry is a logrus.FieldLogger.
-type StructuredLoggerEntry struct {
-	Logger logrus.FieldLogger
-}
-
-func (l *StructuredLoggerEntry) Write(status, bytes int, elapsed time.Duration) {
-	l.Logger = l.Logger.WithFields(logrus.Fields{
-		"resp_status":       status,
-		"resp_bytes_length": bytes,
-		"resp_elapsed_ms":   float64(elapsed.Nanoseconds()) / 1000000.0,
-	})
-
-	l.Logger.Infoln("request complete")
-}
-
-// Panic prints stack trace
-func (l *StructuredLoggerEntry) Panic(v interface{}, stack []byte) {
-	l.Logger = l.Logger.WithFields(logrus.Fields{
-		"stack": string(stack),
-		"panic": fmt.Sprintf("%+v", v),
-	})
-}
-
-// Helper methods used by the application to get the request-scoped
-// logger entry and set additional fields between handlers.
-
-// GetLogEntry return the request scoped logrus.FieldLogger.
-//noinspection GoUnusedExportedFunction
-func GetLogEntry(r *http.Request) logrus.FieldLogger {
-	entry := middleware.GetLogEntry(r).(*StructuredLoggerEntry)
-	return entry.Logger
-}
-
-// LogEntrySetField adds a field to the request scoped logrus.FieldLogger.
-//noinspection GoUnusedExportedFunction
-func LogEntrySetField(r *http.Request, key string, value interface{}) {
-	if entry, ok := r.Context().Value(middleware.LogEntryCtxKey).(*StructuredLoggerEntry); ok {
-		entry.Logger = entry.Logger.WithField(key, value)
-	}
-}
-
-// LogEntrySetFields adds multiple fields to the request scoped logrus.FieldLogger.
-//noinspection GoUnusedExportedFunction
-func LogEntrySetFields(r *http.Request, fields map[string]interface{}) {
-	if entry, ok := r.Context().Value(middleware.LogEntryCtxKey).(*StructuredLoggerEntry); ok {
-		entry.Logger = entry.Logger.WithFields(fields)
-	}
-}
