+++ conf/default.yaml
@@ -11,7 +11,7 @@ database:
     parseTime: true # required for the migration system
 logging:
   textlogging: true
-  logLevel: debug # debug, info, warning, error, fatal, panic
+  level: debug # debug, info, warning, error, fatal, panic
 
 # reverseProxy: # server to which requests which do not match any route will be sent
 #   server: "http://localhost:3000"
